Move the covariate functions to their own module
