+++ src/harpstrata/Covariate/deduceHarpKeyId/index.ts
@@ -1,7 +1,7 @@
-import type { PitchIds } from '../types'
-import { getOrderedPitchIds } from '../pitchMap'
 import { getPozitionRootOffset } from '../../Pozition'
 import type { PozitionIds } from '../../Pozition'
+import { getOrderedPitchIds } from '../../Pitch'
+import type { PitchIds } from '../../Pitch'
 
 
 type DeduceKeyPitchProps = {
