+++ src/harpstrata/Covariate/index.ts
@@ -1,0 +1,3 @@
+export { deduceHarpKeyId } from './deduceHarpKeyId'
+export { deduceRootPitchId } from './deduceRootPitchId'
+export { deducePozitionId } from './deducePozitionId'
