+++ src/harpstrata/Pitch/index.ts
@@ -2,8 +2,6 @@ export { getOrderedPitchIds } from './pitchMap'
 export { getPitch } from './pitchMap'
 
 export { getPitchMatrix } from './getPitchMatrix'
-export { deduceHarpKeyId } from './deduceHarpKeyId'
-export { deduceRootPitchId } from './deduceRootPitchId'
 
 export type { Pitch, PitchRow, PitchMatrix } from './types'
 export { PitchIds } from './types'
