+++ src/harpstrata/Pozition/index.ts
@@ -1,5 +1,4 @@
 export { getOrderedPozitionIds, getPozitionRootOffset, getPozition } from './pozitionMap'
-export { deducePozitionId } from './deducePozitionId'
 
 export { PozitionIds } from './types'
 export type { Pozition } from './types'
